backend/requirements.txt
# =============================================================================
# Backend SaaS Dependencies
# =============================================================================

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Security
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Cache & Queue
redis>=5.0.0

# Billing
httpx>=0.25.2

# Environment
python-dotenv>=1.0.0

# Observability
sentry-sdk[fastapi]>=1.38.0

# =============================================================================
# Streamlit & AI Agents Dependencies
# =============================================================================

# Streamlit UI
streamlit>=1.28.0

# AI/LLM
anthropic>=0.78.0

# Export formats
reportlab>=4.0.7
python-pptx>=0.6.21
jinja2>=3.1.2

# Data processing
pandas>=2.1.3
openpyxl>=3.1.2
PyPDF2>=3.0.1
tabulate>=0.9.0
